refactor(reminder): revised to match with Session Attributes

+++ backend/src/main/java/com/appdev/academeet/dto/ReminderRequest.java
@@ -4,39 +4,40 @@ import java.time.LocalDateTime;
 
 public class ReminderRequest {
     private Long studentId;
-    private Integer sessionId;
+    private Long sessionId; // UPDATED: Was Integer
     private LocalDateTime reminderTime;
-    
+   
     // Constructors
     public ReminderRequest() {}
-    
-    public ReminderRequest(Long studentId, Integer sessionId, LocalDateTime reminderTime) {
+   
+    // UPDATED: Was Integer
+    public ReminderRequest(Long studentId, Long sessionId, LocalDateTime reminderTime) {
         this.studentId = studentId;
         this.sessionId = sessionId;
         this.reminderTime = reminderTime;
     }
-    
+   
     // Getters and Setters
     public Long getStudentId() {
         return studentId;
     }
-    
+   
     public void setStudentId(Long studentId) {
         this.studentId = studentId;
     }
-    
-    public Integer getSessionId() {
+   
+    public Long getSessionId() { // UPDATED: Was Integer
         return sessionId;
     }
-    
-    public void setSessionId(Integer sessionId) {
+   
+    public void setSessionId(Long sessionId) { // UPDATED: Was Integer
         this.sessionId = sessionId;
     }
-    
+   
     public LocalDateTime getReminderTime() {
         return reminderTime;
     }
-    
+   
     public void setReminderTime(LocalDateTime reminderTime) {
         this.reminderTime = reminderTime;
     }
